Move `ClientEvent` and `handle_client_event` to eventloop.rs
