+++ crates/spartan-farmer/src/dht/client.rs
@@ -1,23 +1,13 @@
-// Stuff for Kademlia
-use libp2p::kad::{QueryId, QueryResult};
-
-// Stuff needed to create the swarm
-use libp2p::{Multiaddr, PeerId};
-
-// Stuff needed to set up channels between Client API task and EventLoop task.
+use super::{
+    core::create_node,
+    eventloop::{ClientEvent, EventLoop},
+};
 use futures::channel::{
     mpsc::{channel, Sender},
     oneshot,
 };
 use futures::prelude::*;
-
-type OneshotError = Box<dyn std::error::Error + Send>;
-type OneshotType = Result<(), OneshotError>;
-
-use super::{
-    core::create_node,
-    eventloop::{ClientEvent, EventLoop},
-};
+use libp2p::{kad::QueryId, Multiaddr, PeerId};
 
 pub struct ClientConfig {
     pub bootstrap_nodes: Vec<String>, // Vec<(Multiaddr, PeerId)>,
@@ -133,145 +123,11 @@ impl Client {
     }
 }
 
-pub(super) enum ClientEvent {
-    // Event for adding a listening address.
-    Listen {
-        addr: Multiaddr,
-        sender: oneshot::Sender<OneshotType>,
-    },
-    // List all known peers.
-    // TODO: We are using for testing.
-    #[allow(dead_code)]
-    KnownPeers {
-        sender: oneshot::Sender<Vec<PeerId>>,
-    },
-    // Dial another peer.
-    // TODO: We are using for testing.
-    #[allow(dead_code)]
-    Dial {
-        addr: Multiaddr,
-        peer: PeerId,
-        sender: oneshot::Sender<OneshotType>,
-    },
-    // Bootstrap during the initial connection to the DHT.
-    // NOTE: All the bootstrap nodes must already be connected to the swarm before we can start the
-    // bootstrap process.
-    Bootstrap {
-        sender: oneshot::Sender<QueryId>,
-    },
-    // Get all listening addresses.
-    // TODO: We are using for testing.
-    #[allow(dead_code)]
-    Listeners {
-        sender: oneshot::Sender<Vec<Multiaddr>>,
-    },
-    // Read Kademlia Query Result.
-    // TODO: We are using for testing.
-    #[allow(dead_code)]
-    QueryResult {
-        qid: QueryId,
-        sender: oneshot::Sender<String>,
-    },
-}
-
-pub(super) fn handle_client_event(eventloop: &mut EventLoop, event: ClientEvent) {
-    match event {
-        ClientEvent::Listen { addr, sender } => match eventloop.swarm.listen_on(addr) {
-            Ok(_) => sender.send(Ok(())).unwrap(),
-            Err(e) => sender.send(Err(Box::new(e))).unwrap(),
-        },
-        ClientEvent::Bootstrap { sender } => {
-            if let Ok(qid) = eventloop.swarm.behaviour_mut().kademlia.bootstrap() {
-                sender.send(qid).unwrap();
-            }
-        }
-        ClientEvent::KnownPeers { sender } => {
-            let mut result = Vec::new();
-
-            for bucket in eventloop.swarm.behaviour_mut().kademlia.kbuckets() {
-                for record in bucket.iter() {
-                    result.push(*record.node.key.preimage());
-                }
-            }
-
-            sender.send(result).unwrap();
-        }
-        ClientEvent::Dial { addr, peer, sender } => {
-            eventloop
-                .swarm
-                .behaviour_mut()
-                .kademlia
-                .add_address(&peer, addr.clone());
-
-            eventloop.swarm.dial_addr(addr).unwrap();
-
-            sender.send(Ok(())).unwrap();
-        }
-        ClientEvent::Listeners { sender } => {
-            sender
-                .send(eventloop.swarm.listeners().cloned().collect::<Vec<_>>())
-                .unwrap();
-        }
-        ClientEvent::QueryResult { qid, sender } => {
-            if eventloop.query_result.contains_key(&qid) {
-                let result = match eventloop.query_result.remove(&qid).unwrap() {
-                    QueryResult::Bootstrap(result) => match result {
-                        Ok(result) => format!(
-                            "[RESULT] This query still has {:?} peers remaining.",
-                            result.num_remaining
-                        ),
-                        Err(e) => format!("{:?}", e),
-                    },
-                    QueryResult::GetClosestPeers(result) => match result {
-                        Ok(result) => {
-                            format!("This query produced {:?} peers.", result.peers.len())
-                        }
-                        Err(e) => format!("{:?}", e),
-                    },
-                    _ => "Unknown QueryResult Type".to_string(),
-                };
-
-                sender.send(result).unwrap();
-            } else {
-                let query = eventloop
-                    .swarm
-                    .behaviour_mut()
-                    .kademlia
-                    .query(&qid)
-                    .unwrap();
-
-                let stats = format!(
-                    "Total Requests: {}\nFailed: {}\nSucceded: {}\nPending: {}\n",
-                    query.stats().num_requests(),
-                    query.stats().num_failures(),
-                    query.stats().num_successes(),
-                    query.stats().num_pending()
-                );
-
-                let info = match query.info() {
-                    QueryInfo::Bootstrap { remaining, .. } => {
-                        format!(
-                            "[INFO] This query still has {:?} peers remaining.",
-                            remaining
-                        )
-                    }
-                    _ => "Unknown QueryInfo Type".to_string(),
-                };
-
-                sender.send(stats + &info).unwrap();
-            }
-        }
-    }
-}
-
 // This method will construct a new Swarm and EventLoop object.
 pub fn create_connection(config: &ClientConfig) -> (Client, EventLoop) {
     let (client_tx, client_rx) = channel(10);
 
-    let (peerid, swarm) = match config.client_type {
-        ClientType::Bootstrap => create_node(config),
-        ClientType::Normal => create_node(config),
-    };
+    let (peerid, swarm) = create_node(config);
 
     let eventloop = EventLoop::new(swarm, client_rx);
     let client = Client::new(peerid, client_tx);
